Reparado el hamburguer de administrador

+++ app/src/main/java/com/tute/estacionamientofrlp/CompraActivity.java
@@ -299,7 +299,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
 
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Default));}
 
@@ -315,7 +315,7 @@ public class CompraActivity extends AppCompatActivity implements Serializable, S
                     Log.e("ERROR3.2", String.valueOf(pateComp));
                     montoacomp = (pateComp.size()*Constantes.precioticket);
                     tv3.setText("Costo: $"+ montoacomp);
-                    if (montoacomp > Integer.parseInt(saldog)) {
+                    if (montoacomp > Double.parseDouble(saldog)) {
                         tv3.setTextColor(getResources().getColor(R.color.Red));
                     } else {tv3.setTextColor(getResources().getColor(R.color.Default));}
                     c6.setChecked(false);
